+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## dev-master
 
+* Added `--wait` option to command `fortune`.
 * Restructured import data.
 * Added strict typing.
 * Added `--short` option to command `fortune`.
@@ -10,7 +11,7 @@
 * Added command `index`.
 * Added `--length` option to command `fortune`. 
 * Added `--author` option to command `fortune`. 
-* Set the default `--wordwrap` og command `fortune` to the terminal window width minus one.
+* Set the default `--wordwrap` to command `fortune` to the terminal window width minus one.
 
 ## 1.0.0 - 2018-xx-xx